add function for delete items and set popup dialouge box for confirmation and est it

+++ src/pages/dataInput/AddCharges.js
@@ -75,16 +75,21 @@ const AddCharges = () => {
   // http://localhost:5001/delete/:id
 
   const handleDelete = async (id) => {
-    try {
-      await axios.delete(
-        `https://grozziie.zjweiting.com:3091/web-api-tht-1/api/dev/addcharges/${id}`
-      );
-      toast.warn("Data successfully Deleted!!", { position: "top-center" });
-      fetchAccounts();
-    } catch (error) {
-      toast.error("You can't delete now. Please try again later!", {
-        position: "top-center",
-      });
+    const confirmDelete = window.confirm(
+      "Are you sure, you want to delete this Charge?"
+    );
+    if (confirmDelete) {
+      try {
+        await axios.delete(
+          `https://grozziie.zjweiting.com:3091/web-api-tht-1/api/dev/addcharges/${id}`
+        );
+        toast.warn("Data successfully Deleted!!", { position: "top-center" });
+        fetchAccounts();
+      } catch (error) {
+        toast.error("You can't delete now. Please try again later!", {
+          position: "top-center",
+        });
+      }
     }
   };
